Honor the host specified in current docker context
One can have multiple docker hosts that they manage and can switch
between them using the "docker context" command.
Previously, we used to connect to either the "default docker host" (for
example, unix:///var/run/docker.sock in case of Linux/macOS) or to the
host specified via the "DOCKER_HOST" environment variable.
However, if one switched to a different context, we had no way of detecting
that.

Now, we determine the docker host that we connect to by evaluating the
following sources in decreasing order of precedence:
- value of "DOCKER_HOST" environment variable
- host specified in the context set via the "DOCKER_CONTEXT" environment variable
- "default docker host" as applicable, otherwise

Fixes: #285

Signed-off-by: Rajiv Kushwaha <raj25by10@gmail.com>